fix: do not mutate input options
