+++ src/legacy_wrappers/mongo_client.js
@@ -1,6 +1,8 @@
 'use strict';
 
-const { toLegacy, maybeCallback, addLegacyMetadata } = require('../utils');
+const { toLegacy, maybeCallback } = require('../utils');
+
+const { version } = require('../../package.json');
 
 module.exports = Object.create(null);
 Object.defineProperty(module.exports, '__esModule', { value: true });
@@ -12,7 +14,37 @@ module.exports.makeLegacyMongoClient = function (baseClass) {
       if (options == null) {
         options = {};
       }
-      addLegacyMetadata(options);
+
+      const incorrectOptionsType = typeof options !== 'object';
+      const incorrectDriverInfo =
+        options.driverInfo != null && typeof options.driverInfo !== 'object';
+      if (incorrectOptionsType || incorrectDriverInfo) {
+        // Pass this mistake along to the MongoClient constructor
+        super(connectionString, options);
+        return;
+      }
+
+      options = {
+        ...options,
+        driverInfo: options.driverInfo == null ? {} : { ...options.driverInfo }
+      };
+
+      const infoParts = {
+        name: ['mongodb-legacy'],
+        version: [version]
+      };
+
+      // name handling
+      if (typeof options.driverInfo.name === 'string') {
+        infoParts.name.push(options.driverInfo.name);
+      }
+      options.driverInfo.name = infoParts.name.join('|');
+
+      // version handling
+      if (typeof options.driverInfo.version === 'string') {
+        infoParts.version.push(options.driverInfo.version);
+      }
+      options.driverInfo.version = infoParts.version.join('|');
 
       super(connectionString, options);
     }
